Refactored into separate controllers. Some auto layout constraints pending.
