+++ WBooks.xcodeproj/xcuserdata/gmaspero.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -10,13 +10,13 @@
             ignoreCount = "0"
             continueAfterRunningActions = "No"
             filePath = "WBooks/Library/MainMenuController.swift"
-            timestampString = "581378329.253293"
+            timestampString = "581554247.570819"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "67"
-            endingLineNumber = "67"
-            landmarkName = "MainMenuController"
-            landmarkType = "21">
+            startingLineNumber = "68"
+            endingLineNumber = "68"
+            landmarkName = "tableView(_:numberOfRowsInSection:)"
+            landmarkType = "7">
          </BreakpointContent>
       </BreakpointProxy>
    </Breakpoints>
