+++ WBooks/Book Info/BookInfoView.swift
@@ -1,0 +1,15 @@
+//
+//  DetailView.swift
+//  WBooks
+//
+//  Created by Gaston Maspero on 30/05/2019.
+//  Copyright © 2019 Wolox. All rights reserved.
+//
+
+import UIKit
+import WolmoCore
+
+class BookInfoView: UIView, NibLoadable {
+    
+
+}
